v5 - check if there is a command context

+++ modules/activiti5-engine/src/main/java/org/activiti5/engine/impl/pvm/runtime/AtomicOperationTransitionDestroyScope.java
@@ -15,6 +15,7 @@ package org.activiti5.engine.impl.pvm.runtime;
 import java.util.List;
 
 import org.activiti5.engine.impl.context.Context;
+import org.activiti5.engine.impl.interceptor.CommandContext;
 import org.activiti5.engine.impl.persistence.entity.JobEntity;
 import org.activiti5.engine.impl.pvm.process.ActivityImpl;
 import org.activiti5.engine.impl.pvm.process.ScopeImpl;
@@ -91,13 +92,18 @@ public class AtomicOperationTransitionDestroyScope implements AtomicOperation {
 
         // TODO!
         execution.destroy();
+        
         // GDH - fix for ACT-4263
-        List<JobEntity> jobs = Context.getCommandContext()
-        		.getJobEntityManager()
-        		.findJobsByExecutionId(execution.getId());
-        for (JobEntity job: jobs) {
-        	job.delete();
-        }		
+        CommandContext commandContext = Context.getCommandContext();
+        if (commandContext != null) {
+            List<JobEntity> jobs = Context.getCommandContext()
+            		.getJobEntityManager()
+            		.findJobsByExecutionId(execution.getId());
+            for (JobEntity job: jobs) {
+            	job.delete();
+            }		
+        }
+        
         propagatingExecution = execution;
         
       } else {
